UPdate send order

+++ app/Http/Controllers/QueueOrderController.php
@@ -81,6 +81,12 @@ class QueueOrderController extends Controller
     public function cancelQueueOrder()
     {
         $param = request()->all();
+        $validator = Validator::make($param, [
+            'table_id' => 'required'
+        ]);
+        if ($validator->fails()) {
+            return $this->errorResponse($validator->errors(), null, false, 404);
+        }
         $this->queueOrderService->delete($param['id']);
 
         return $this->successResponse(null, 'Delete Success');
